requirements.txt
Flask==3.0.0
web3==6.11.3
py-solc-x==2.0.2
markupsafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
